+++ vagrant/etc/profile.d
@@ -1,0 +1,1 @@
+../guest/etc/profile.d