+++ vagrant/lib/provision.rb
@@ -17,7 +17,7 @@ def bootstrap_sh (conf, roles, env = {})
       base_dir: BASE_DIR,
       vagrant_dir: VAGRANT_DIR,
       shared_dir: SHARED_DIR,
-      allowable_roles: ENV['VAGRANT_ALLOWABLE_ROLES'],
+      ssl_dir: SHARED_DIR + '/ssl',
       bootstrap_log: '/var/log/bootstrap.log',
       host_zoneinfo: File.readlink('/etc/localtime')
     }.merge(env)
@@ -33,30 +33,7 @@ def bootstrap_sh (conf, roles, env = {})
   end
 end
 
-# Performs a service call on the guest
-# Params:
-# +conf+:: vagrant provisioning conf object
-# +calls+:: Hash key/value pairs where key is the verb and value is an array of services the verb will act on
-def service (conf, calls)
-  service_sh = ""
-
-  calls.each do | key, val |
-    if not val.is_a?(Array)
-      val = [val]
-    end
-    
-    val.each do | val |
-      service_sh = "#{service_sh}\nservice #{val} #{key} 2> >(grep -v -f #{VAGRANT_DIR}/etc/filters/service >&2)"
-    end
-  end
-
-  conf.vm.provision :shell, run: 'always' do |conf|
-    conf.name = "service_sh"
-    conf.inline = service_sh
-  end
-end
-
-# Configure the machines memory allocation
+# Configure the guest memory allocation
 # Params:
 # +conf+:: vagrant provisioning conf object
 # +ram+:: amount of memory specified in megabytes
@@ -65,7 +42,7 @@ def vm_set_ram (conf, ram)
   conf.vm.provider('vmware_fusion') { |vm| vm.vmx['memsize'] = ram; }
 end
 
-# Configure the machines CPU allocation
+# Configure the guest CPU allocation
 # Params:
 # +conf+:: vagrant provisioning conf object
 # +cpu+:: number of CPUs to allocate