+++ vagrant/scripts/solr.sh
@@ -1,0 +1,14 @@
+#!/usr/bin/env bash
+##
+ # Copyright © 2016 by David Alger. All rights reserved
+ # 
+ # Licensed under the Open Software License 3.0 (OSL-3.0)
+ # See included LICENSE file for full text of OSL-3.0
+ # 
+ # http://davidalger.com/contact/
+ ##
+
+set -e
+
+# TODO: install and configure java
+# TODO: install and configure solr service